requirements.txt
# Claude SDK
anthropic>=0.40.0

# Atlassian API Client
atlassian-python-api>=3.41.0

# HTTP Client
httpx>=0.27.0

# Environment Management
python-dotenv>=1.0.0

# JSON handling
pydantic>=2.0.0
